release 3.1.95 source code

+++ Services/CodeArtsPipeline/V2/Model/ShowStepOutputsResponse.php
@@ -40,7 +40,7 @@ class ShowStepOutputsResponse implements ModelInterface, ArrayAccess
     */
     protected static $openAPIFormats = [
         'stepOutputs' => null,
-        'currentSystemTime' => 'int32'
+        'currentSystemTime' => 'int64'
     ];
 
     /**
